refactoring [Try DEMO] on Main Banner
